pyproject.toml
[project]
name = "az-ai-catalyst"
version = "0.1.0"
description = "AZ AI Catalyst : multi-agentic & reasoning document processing"
readme = "README.md"
authors = [
    { name = "Dominique Broeglin", email = "dominique.broeglin@microsoft.com" },
    { name = "Evgeny Minkevich", email = "evgeny.minkevich@microsoft.com" }
]
requires-python = ">=3.12"

dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-inference[opentelemetry]>=1.0.0b9",
    "azure-ai-projects>=1.0.0b10",
    "azure-identity>=1.21.0",
    "azure-search-documents>=11.5.2",
    "mlflow>=2.22.0",
    "openai>=1.76.0",
    "pedantic>=2.1.9",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pymupdf>=1.25.5",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "semantic-text-splitter>=0.25.1",
    "typer>=0.15.2",
]

[project.scripts]
az-ai-catalyst = "az_ai.catalyst:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["az_ai"]

[tool.az_ai.catalyst]
#model_name = "gpt-4.1-mini-2025-04-14"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = [
    "--tb=short",
    "--disable-warnings",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.7",
]

[tool.ruff]
line-length = 120
fix = true
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
extend-fixable = ["E", "F", "UP", "B", "SIM", "I"]
